Problem Statement.
Problem Statement:
 In this Contest case study we work on data related to machine breakdowns
to predict future occurrences and create a framework for preventive maintenance.
The above machines work on the principal of pressure at 3 points. The company wants to build a
model to predict breakdown and use preventive maintenance to reduce downtime (time during which
production is stopped especially during setup for an operation or when making repairs).